requirements.txt
Flask>=3.0.0
requests>=2.32.0
pydantic>=2.8.0
python-dotenv>=1.0.1
pytest>=8.2.0
